refactor: move initialize tests into separate file
