+++ e-token/tests/execute_transfer_callback.rs
@@ -1,13 +1,14 @@
 use crate::common::callback_mock::take_execute_callbacks;
-use crate::common::crank_mock::take_execute_cranks;
 use crate::common::magic_mock::{take_captured_ephemeral_closes, take_captured_ephemeral_creates};
-use crate::common::{callback_mock, crank_mock, magic_mock};
+use crate::common::{callback_mock, magic_mock};
+use ephemeral_rollups_pinocchio::consts::MAGIC_PROGRAM_ID;
 use ephemeral_spl_api::state::group_receipt::{GroupReceipt, GroupReceiptHeader};
 use ephemeral_spl_api::state::transfer_queue::{HEADER_LEN, QUEUE_SEED, TRANSFER_QUEUE_VERSION};
+use ephemeral_spl_api::ID as PROGRAM;
 use ephemeral_token_program::TransferCallbackArgs;
-use magicblock_magic_program_api::instruction::{CallbackInstruction, MagicBlockInstruction};
-use magicblock_magic_program_api::pda::{CALLBACK_SIGNER, CRANK_SIGNER};
-use magicblock_magic_program_api::{CALLBACK_PROGRAM_ID, CRANK_PROGRAM_ID};
+use magicblock_magic_program_api::instruction::CallbackInstruction;
+use magicblock_magic_program_api::pda::CALLBACK_SIGNER;
+use magicblock_magic_program_api::CALLBACK_PROGRAM_ID;
 use serial_test::serial;
 use solana_account::Account as SolanaAccount;
 use solana_instruction::{AccountMeta, Instruction};
@@ -22,8 +23,6 @@ use utils::TestInternalInstruction as internal;
 mod common;
 mod utils;
 
-const PROGRAM: Pubkey = pubkey!("SPLxh1LVZzEkX99H6rqYizhytLWPZVV296zyYDPagv2");
-const MAGIC_PROGRAM: Pubkey = pubkey!("Magic11111111111111111111111111111111111111");
 const MAGIC_VAULT: Pubkey = pubkey!("MagicVau1t999999999999999999999999999999999");
 
 const GROUP_RECEIPT_SEED: &[u8] = b"group-receipt";
@@ -62,16 +61,6 @@ fn callback_ix_data(ok: bool, amount: u64, group_id: u32) -> Vec<u8> {
     data
 }
 
-/// Serialise the initialize_group_receipt instruction data.
-/// Layout: discriminator(1) + group_id(4) + splits(4)
-fn initialize_group_receipt_ix_data(group_id: u32, splits: u32) -> Vec<u8> {
-    let mut data = Vec::with_capacity(9);
-    data.push(internal::InitializeGroupReceipt.discriminator());
-    data.extend_from_slice(&group_id.to_le_bytes());
-    data.extend_from_slice(&splits.to_le_bytes());
-    data
-}
-
 fn derive_queue(mint: Pubkey, validator: Pubkey) -> (Pubkey, u8) {
     Pubkey::find_program_address(&[QUEUE_SEED, mint.as_ref(), validator.as_ref()], &PROGRAM)
 }
@@ -107,17 +96,6 @@ fn receipt_account_data(group_id: u32, splits: u32, bump: u8) -> Vec<u8> {
     data
 }
 
-/// Build a receipt buffer with `splits=0` (partially initialized) and enough
-/// space for `capacity` items. Used to test the "already exists, set splits"
-/// path of process_initialize_group_receipt.
-fn receipt_account_data_partial(group_id: u32, bump: u8, capacity: u32) -> Vec<u8> {
-    // splits=0 means not yet fully initialized
-    let mut data = vec![0u8; GroupReceipt::required_size(capacity as usize)];
-    let header = GroupReceiptHeader::new(group_id, bump, 0);
-    data[..GroupReceiptHeader::size()].copy_from_slice(bytemuck::bytes_of(&header));
-    data
-}
-
 /// Common fixture: a ProgramTest context with queue and receipt accounts
 /// pre-populated. Returns (context, validator keypair, mint, queue, receipt).
 async fn setup_context(
@@ -146,7 +124,6 @@ async fn setup_context(
     pt.add_program("ephemeral_token_program", PROGRAM, None);
     magic_mock::add_mock(&mut pt);
     callback_mock::add_mock(&mut pt);
-    crank_mock::add_mock(&mut pt);
 
     let queue_data = queue_account_data(mint, validator.pubkey(), queue_bump);
     pt.add_account(
@@ -197,7 +174,7 @@ async fn setup_context(
 
     let ctx = pt.start_with_context().await;
 
-    magic_mock::clear_all_captured(MAGIC_PROGRAM);
+    magic_mock::clear_all_captured(MAGIC_PROGRAM_ID);
 
     (ctx, validator, mint, queue, receipt, vault, vault_token)
 }
@@ -268,59 +245,10 @@ fn callback_ix(
             AccountMeta::new_readonly(solana_system_interface::program::ID, false),
             AccountMeta::new_readonly(spl_token_interface::ID, false),
             AccountMeta::new(MAGIC_VAULT, false),
-            AccountMeta::new_readonly(MAGIC_PROGRAM, false),
+            AccountMeta::new_readonly(MAGIC_PROGRAM_ID, false),
         ],
         data: callback_ix_data(ok, amount, group_id),
     }
-}
-
-fn initialize_group_receipt_ix(
-    queue: Pubkey,
-    receipt: Pubkey,
-    group_id: u32,
-    splits: u32,
-) -> Instruction {
-    Instruction {
-        program_id: PROGRAM,
-        accounts: vec![
-            AccountMeta::new(CRANK_SIGNER, true),
-            AccountMeta::new(queue, false),
-            AccountMeta::new(receipt, false),
-            AccountMeta::new(MAGIC_VAULT, false),
-            AccountMeta::new_readonly(MAGIC_PROGRAM, false),
-        ],
-        data: initialize_group_receipt_ix_data(group_id, splits),
-    }
-}
-
-fn crank_executor_ix(
-    validator: Pubkey,
-    queue: Pubkey,
-    receipt: Pubkey,
-    group_id: u32,
-    splits: u32,
-) -> Instruction {
-    let inner_ix = initialize_group_receipt_ix(queue, receipt, group_id, splits);
-
-    let mut account_metas = vec![
-        AccountMeta::new_readonly(validator, true),
-        AccountMeta::new_readonly(CRANK_SIGNER, false),
-        AccountMeta::new_readonly(inner_ix.program_id, false),
-    ];
-    account_metas.extend(inner_ix.accounts.clone().into_iter().map(|mut el| {
-        // CRANK_SIGNER may be set to true in inner_instruction
-        // Outer instruction can't have PDA as signer
-        el.is_signer = false;
-        el
-    }));
-
-    Instruction::new_with_bincode(
-        CRANK_PROGRAM_ID,
-        &MagicBlockInstruction::ExecuteCrank {
-            instructions: vec![inner_ix],
-        },
-        account_metas,
-    )
 }
 
 // ── tests ─────────────────────────────────────────────────────────────────────
@@ -366,11 +294,11 @@ async fn execute_callback_with_pre_initialized_receipt_no_magic_cpi() {
     assert_eq!(executed_callbacks.len(), 1);
 
     // No CreateEphemeralAccount should have been called — receipt already existed.
-    let creates = take_captured_ephemeral_creates(MAGIC_PROGRAM);
+    let creates = take_captured_ephemeral_creates(MAGIC_PROGRAM_ID);
     assert!(creates.is_empty(), "expected no CreateEphemeralAccount CPI");
 
     // No CloseEphemeralAccount either — this is not the last transfer.
-    let closes = take_captured_ephemeral_closes(MAGIC_PROGRAM);
+    let closes = take_captured_ephemeral_closes(MAGIC_PROGRAM_ID);
     assert!(closes.is_empty(), "expected no CloseEphemeralAccount CPI");
 
     // Receipt transfer_completed should be 1.
@@ -425,141 +353,17 @@ async fn execute_callback_closes_receipt_when_last_transfer_with_pre_initialized
     res.result.unwrap();
 
     // No CreateEphemeralAccount — receipt was pre-initialized.
-    let creates = magic_mock::take_captured_ephemeral_creates(MAGIC_PROGRAM);
+    let creates = magic_mock::take_captured_ephemeral_creates(MAGIC_PROGRAM_ID);
     assert!(creates.is_empty(), "expected no CreateEphemeralAccount CPI");
 
     // CloseEphemeralAccount must have been called once — this was the last transfer.
-    let closes = magic_mock::take_captured_ephemeral_closes(MAGIC_PROGRAM);
+    let closes = magic_mock::take_captured_ephemeral_closes(MAGIC_PROGRAM_ID);
     assert_eq!(
         closes.len(),
         1,
         "expected exactly one CloseEphemeralAccount CPI"
     );
-
-    // Log confirming all transfers complete.
-    let logs = res
-        .metadata
-        .as_ref()
-        .map(|m| &m.log_messages)
-        .expect("expected log messages");
-    assert!(
-        logs.iter()
-            .any(|l| l.contains("All transfers complete for group id")),
-        "expected 'All transfers complete' log; got:\n{}",
-        logs.join("\n"),
-    );
 }
-
-/// process_initialize_group_receipt: receipt account is pre-created in the
-/// test context (simulating what the magic program would have allocated).
-/// The processor takes the "already owned by program" path, sets splits,
-/// and does not close because no transfers have been recorded yet.
-#[tokio::test]
-#[serial]
-async fn initialize_group_receipt_sets_splits_on_existing_receipt() {
-    let group_id: u32 = 42;
-    let splits: u32 = 3;
-
-    // Pre-create with splits=0 (partially initialised) and enough capacity for `splits` items.
-    let receipt_data = receipt_account_data_partial(group_id, 0, splits);
-    let (ctx, validator, _mint, queue, receipt, _vault, _vault_token) =
-        setup_context(receipt_data, group_id).await;
-
-    let ix = crank_executor_ix(validator.pubkey(), queue, receipt, group_id, splits);
-
-    let tx = Transaction::new_signed_with_payer(
-        &[ix],
-        Some(&validator.pubkey()),
-        &[&validator],
-        ctx.last_blockhash,
-    );
-    ctx.banks_client.process_transaction(tx).await.unwrap();
-
-    let executed_cranks = take_execute_cranks();
-    assert_eq!(executed_cranks.len(), 1);
-
-    // No close — no transfers have been recorded.
-    let closes = common::magic_mock::take_captured_ephemeral_closes(MAGIC_PROGRAM);
-    assert!(closes.is_empty(), "expected no CloseEphemeralAccount CPI");
-
-    // Receipt must now have splits set correctly.
-    let account = ctx
-        .banks_client
-        .get_account(receipt)
-        .await
-        .unwrap()
-        .expect("receipt must still exist");
-    let header =
-        bytemuck::try_from_bytes::<GroupReceiptHeader>(&account.data[..GroupReceiptHeader::size()])
-            .unwrap();
-    assert_eq!(header.splits(), splits);
-    assert_eq!(header.transfer_completed(), 0);
-}
-
-/// process_initialize_group_receipt: receipt pre-created with splits=0 and
-/// one transfer already recorded. When initialized with splits=1, the program
-/// detects all callbacks are done and CPIs CloseEphemeralAccount.
-#[tokio::test]
-#[serial]
-async fn initialize_group_receipt_closes_when_all_callbacks_already_done() {
-    use ephemeral_spl_api::state::group_receipt::TransferReceipt;
-
-    let group_id: u32 = 7;
-    let splits: u32 = 1;
-
-    // Build a receipt that has 1 transfer recorded but splits still 0.
-    let mut receipt_data = receipt_account_data_partial(group_id, 0, splits);
-    // Manually write transfers_completed = 1 into the header.
-    // GroupReceiptHeader layout: id(4) + splits(4) + transfers_completed(4) + bump(1) + _reserved(11)
-    // transfers_completed is at offset 8.
-    receipt_data[8..12].copy_from_slice(&1u32.to_le_bytes());
-    // Write one dummy TransferReceipt item after the header.
-    let item = TransferReceipt::new(None, 99, true);
-    let header_size = GroupReceiptHeader::size();
-    receipt_data[header_size..header_size + TransferReceipt::size()]
-        .copy_from_slice(bytemuck::bytes_of(&item));
-
-    let (ctx, validator, _mint, queue, receipt, _vault, _vault_token) =
-        setup_context(receipt_data, group_id).await;
-
-    let ix = crank_executor_ix(validator.pubkey(), queue, receipt, group_id, splits);
-
-    let tx = Transaction::new_signed_with_payer(
-        &[ix],
-        Some(&validator.pubkey()),
-        &[&validator],
-        ctx.last_blockhash,
-    );
-
-    let res = ctx
-        .banks_client
-        .process_transaction_with_metadata(tx)
-        .await
-        .unwrap();
-    res.result.unwrap();
-
-    let executed_cranks = take_execute_cranks();
-    assert_eq!(executed_cranks.len(), 1);
-
-    // CloseEphemeralAccount must be called — all callbacks already done.
-    let closes = common::magic_mock::take_captured_ephemeral_closes(MAGIC_PROGRAM);
-    assert_eq!(closes.len(), 1, "expected one CloseEphemeralAccount CPI");
-
-    // Log confirming all transfers complete.
-    let logs = res
-        .metadata
-        .as_ref()
-        .map(|m| &m.log_messages)
-        .expect("expected log messages");
-    assert!(
-        logs.iter()
-            .any(|l| l.contains("All transfers complete for group id")),
-        "expected 'All transfers complete' log; got:\n{}",
-        logs.join("\n"),
-    );
-}
-
-// ── original tests (kept) ─────────────────────────────────────────────────────
 
 /// Mid-group callback: receipt already exists, not the last transfer.
 /// No CPI to the magic program — only the receipt state changes.
@@ -652,17 +456,6 @@ async fn execute_callback_closes_receipt_on_last_transfer() {
     let executed_callbacks = take_execute_callbacks();
     assert_eq!(executed_callbacks.len(), 1);
 
-    // The program emits "All transfers complete for group id: ..." when it reaches the close
-    // path. Verify this log is present — it confirms the right code path was taken.
-    let logs = res
-        .metadata
-        .as_ref()
-        .map(|m| &m.log_messages)
-        .expect("expected log messages in transaction metadata");
-    assert!(
-        logs.iter()
-            .any(|l| l.contains("All transfers complete for group id")),
-        "expected 'All transfers complete' log; got:\n{}",
-        logs.join("\n"),
-    );
+    let closes = take_captured_ephemeral_closes(MAGIC_PROGRAM_ID);
+    assert_eq!(closes.len(), 1);
 }
